Add create_tree_html module to graph package
This commit adds the create_tree_html module to the omisoshiru.graph package. Additionally, it updates the __init__.py file to include the new module in the __all__ list. The test_create_tree_html.py file is also added to the tests/graph directory to provide basic testing for the create_tree_html module.

Changes:
- Added create_tree_html module
- Updated __init__.py to include create_tree_html in __all__
- Added test_create_tree_html.py for basic testing of create_tree_html module