feat: Update WakachiMatcher initialization parameters
Refactor the initialization parameters of WakachiMatcher to include optional bool parameters
unify_hz and unify_hl. These parameters control whether to unify half-width characters and
characters to lowercase, respectively. The default values are set to None, and the class now
initializes them to False if not provided. This change provides more flexibility when creating
instances of WakachiMatcher.