feat: Enhance WakachiReplacer initialization options
Improve the initialization of WakachiReplacer by adding optional bool parameters unify_hz and
unify_hl. These parameters determine whether to unify half-width characters and characters to
lowercase, respectively. The default values are set to None, and if not provided, the class
initializes them to None. This enhancement allows for more customization when creating instances
of WakachiReplacer.